heroku deploy

+++ database/migrations/2020_08_02_222552_create_orders_table.php
@@ -19,7 +19,7 @@ class CreateOrdersTable extends Migration
             $table->string('server')->nullable();
             $table->string('nickname')->nullable();
             $table->string('email')->nullable();            
-            $table->integer('count')->default(0);
+            $table->double('count')->default(0);
             $table->timestamps();
         });
     }
